+++ CHANGELOG.md
@@ -31,7 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Request a new version of an article that has already been accepted
  - [#166] (https://github.com/CCSDForge/episciences/issues/116): hide "guest" role and doi-settings action   
  - The "status" filter is now dynamically created.
-
+ - Prevent "Add sources files" and "Add the formatted version" buttons JS reactivation
+ - Move "copy editing" section
+ 
 ##Fixed
 - Fixed: inability to delete a volume (RT#145178): reset the volume of the previous version when submitting a new version.
 - Improvements in emailing: see github #117.
@@ -40,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed: not allowing the author to be invited: Invite a reviewer > new reviewer > Invite a new user.
 - Fixed: the attachments to the rating report are not available [RT#145473]
 - Fixed: answer revision request: in some situations, the "Submit" button remains inactive.
+- Fixed: IDE warnings
 
 ## 1.0.24.2 - 2022-02-10
 # Fixed